Activity changed

+++ CarRentalCompany.Domain/Models/ActivityInstance.cs
@@ -2,7 +2,7 @@
 {
     public class ActivityInstance
     {
-        public ActivityInstance(string name, int orderNo, Guid activityDefinitionId, Guid receiptFormId)
+        public ActivityInstance(string name, double orderNo, Guid activityDefinitionId, Guid receiptFormId)
         {
             Id = Guid.NewGuid();
             Name = name;
@@ -13,7 +13,7 @@
             ReceiptFormId = receiptFormId;
         }
 
-        public ActivityInstance(Guid id, string name, string payload, int orderNo, bool isCompleted, Guid activityDefinitionId, Guid receiptFormId) 
+        public ActivityInstance(Guid id, string name, string payload, double orderNo, bool isCompleted, Guid activityDefinitionId, Guid receiptFormId) 
             : this(name, orderNo, activityDefinitionId, receiptFormId)
         {
             Id = id;
@@ -24,7 +24,7 @@
         public Guid Id { get; }
         public string Name { get; }
         public string Payload { get; }
-        public int OrderNo { get; }
+        public double OrderNo { get; }
         public bool IsCompleted { get; }
         public Guid ActivityDefinitionId { get; }
         public Guid ReceiptFormId { get; }
